User quartils to calc number groups we can handle.
* src/trayLatticeDetector/detectLattice.m (createGroupStruct): placed logic to
create the structure of the groups in a separate inner function.
(groupLines): The group change points and therefore the gruops themselves now
depend on differences that are outliers. This is assuming that the differences
between consecutive lines is big if they are in different groups and is small
if they are in the same group.