Add files to handle the correction of lens distortion.
I stole this from Tims code. I added the ability to undistort coordinates.
This means that we can find relative coordinates in the undistorted image and
then identify them in the original image. Original image is always better!!!!

We also started the logic to automatically identify where the picture came
from. What camera and what lens. this needs to grow into code that
automatically relates a distortion constant to a camera, lens and focal
length.