New function getAlternativeIntersects
* src/trayLatticeDetector/detectLattice.m (getAlternativeIntersects): We can't
simply ignore the lines that do not intersect because we might end up with no
intersections at all. To address this we create getAlternativeIntersects. This
function returns the intersection of two 'new' lines made up of the opposing
points of the original lines.
(getLatIntersections): Shortened function name to increase redability.
(Intersection calculations): Fixed erroneous math used to calculated
intersections.