Measure the area and enclosing circle of one leaf.
We assume cropped images of a picture of a back lit leaf and a reference
square.  The square should always be at the bottom left corner and the leaf
should be separated from the square. The result is placed in a csv file.

* src/leafscan/leafscandir.m (leafscandir): This will go through the files and
create a csv file with all the results.
* src/leafscan/leafscanimg.m (leafscanimg): This will detect the square and
leaf. It will measure leaf area and enclosing circle.
* src/leafscan/minboundcircle.m (minboundcircle): Code comming from the Matlab
sharing website. Origin is on the file.
* src/leafscan/parseFilename.m (parseFilename): Takes the name of the file and
extracts important information.This is a generic function and should go
somewhere where we can call it from other places.