New file
First version of the file containing logic that will create rosette masks in
the overall image.

* src/manualRosetteDetect/findSegmentedRosette.m(findSegmentedRosette): This
function will return a mask and an offset where the mask fits into the parent
image.
(getKMeansMask): New function that takes a subimage and analyzes it in the
Excess green space. The idea is to separate the green from the plants from
whatever else is in the subimage.