Adjust rosette return values so we can paint them
* src/manualRosetteDetect/annotation_gui.m(put_image_in_axis): Put image in
handles.img so we can reference and change it in other places.
(segment_Callback): call findSegmentedRosette for all the squares (line
objects) contained in handles.figure1. With the returned offset and mask,
change the image in the axis in order to reflect the detected area.
* src/manualRosetteDetect/findSegmentedRosette.m (findSegmentedRosette): Add
imgOffset as one of the returned arguments. This will contain the offset where
the mask is located within the original image.