Add exceptions and calculate maxGrowth
* src/manualRosetteDetect/findSegmentedRosette.m (findSegmentedRosette): New
variable maxGrowth that defines the maximum size that the evaluating square
can grow to.
(findSegmentedRosette): New variable foundRosette that identifies when we were
successfull at finding a good separation with the k-means.
(findSegmentedRosette): Two new exceptions for when maxGrowth is less than 5
and when we are not able to finde a rosette.