Add area in cm calculation
* src/leafscan/leafscandir.m(leafscandir): Include the radius in cm
calculations in the outpub csv file.
(leafscandir): Assign the value of the radius in cm to R var.
* src/leafscan/leafscanimg.m(leafscanimg): New variable pixRatioLine that will
have the linear ration between pixels and cm. New variables square.pixSide and
square.cmSide that will house the sizes of the square in pixels and cm. Size
of the square side in pixels is the square root of the area of the connected
component that represents the square.
* src/leafscan/showleafscanimg.m(showleafscanimg): Assign the cm radius to R.