Add 'create image series' functionality.
* src/manualRosetteDetect/rosettedetect_gui.m(imgseries_Callback): New method
to catch the call back from new "img Series" button. We ask the user for the
source file where all the images are stored and a destination file where the
new (segmented) image series will go to.
(vector_Callback): New method to handle the callback from the "Vector" button.
Not implemented yet.
* src/manualRosetteDetect/rosettedetect_gui.fig: Add two buttons to the GUI:
"img Series" and "Vector"