Add code to generate area matrix.
* src/manualRosetteDetect/rosettedetect_gui.m(vector_Callback): Saves 3
variables as rosetteAreas.mat. The first variable is ids which contains the id
name of each pot; the second variable is dates which contains all the dates of
the image series and areas which is a matrix that has dates for rows and plant
ids for columns. ids(1) corresponds to areas(:,1) and dates(1) corresponds to
areas(1,:).