Create a more general getKMeansVecMask method.
We can now accept multidimensional data sets.
* src/manualRosetteDetect/getKMeansVecMask.m(getKMeansVecMask): Add the
argument description in the function documentation. We use pdist2 and means to
calculate the distance and mean values of the multidimensional arrays.
* src/manualRosetteDetect/segmentRosette_mask.m(segmentRosette_mask): Modify
mean starting values that are passed to the getKMeansVecMask method.
* src/manualRosetteDetect/segmentRosette_sqr.m(segmentRosette_sqr): Modify the
mean starting value as well as the way we reshape the image before passing
them to the new getKMeansVecMask.