Adjust contour levelset values.
We are adjusting maxiter, growth and perimeter proportion of error. The most
important here being growth as it will fail with a 'big' growth (no contours
are found) but it will also fail with a 'small' growth as it will not get the
growth of the plant.
* src/manualRosetteDetect/segmentRosette_levelset.m(segmentRosette_levelset):
We change the way we calculate the growth. Now it will be 5 or less. We need
to put this variable as a method argument.
(segmentRosette_levelset): We are just doing a mask swelling instead of a
swelling and a dillation. This give a better starting point for the contour
algorithm in the next image.
(segmentRosette_levelset): We throw an error if the mask we detect has more
than 10% 1's in the perimeter.
(activeContours): We calculate all smoothing variables outside the loop.
(activeContours): We use imgradient instead of gradient. It make code more
readable.